folder structuring and a bit of code cleanup
