+++ src/pages/Signup.js
@@ -1,0 +1,10 @@
+import React, { useState } from 'react';
+
+
+export function Signup() {
+    return (
+        <>
+            <h2>Will Be implementing this.</h2>
+        </>
+    );
+}
